+++ e2e/commands/import.e2e.js
@@ -1529,6 +1529,9 @@ console.log(barFoo.default());`;
         helper.importComponent('utils/is-string'); // imports is-type as a dependency
         helper.importComponent('utils/is-type');
         localConsumerFiles = helper.getConsumerFiles();
+
+        const appJsFixture = "const isString = require('./components/utils/is-string'); console.log(isString());";
+        fs.outputFileSync(path.join(helper.localScopePath, 'app.js'), appJsFixture);
       });
       it('should rewrite is-type directly into "components" directory', () => {
         const expectedLocation = path.join('components', 'utils', 'is-type', 'is-type.js');
@@ -1549,10 +1552,17 @@ console.log(barFoo.default());`;
         const isTypeFixtureV2 = "module.exports = function isType() { return 'got is-type v2'; };";
         helper.createFile(path.join('components', 'utils', 'is-type'), 'is-type.js', isTypeFixtureV2);
 
-        const appJsFixture = "const isString = require('./components/utils/is-string'); console.log(isString());";
-        fs.outputFileSync(path.join(helper.localScopePath, 'app.js'), appJsFixture);
         const result = helper.runCmd('node app.js');
         expect(result.trim()).to.equal('got is-type v2 and got is-string');
+      });
+      describe('moving is-type', () => {
+        before(() => {
+          helper.move('components/utils/is-type', 'another-dir');
+        });
+        it('should not break is-string component', () => {
+          const result = helper.runCmd('node app.js');
+          expect(result.trim()).to.have.string('got is-string');
+        });
       });
     });
     describe('import is-type directly and then as a dependency', () => {