+++ src/consumer/bit-map/component-map.js
@@ -110,6 +110,7 @@ export default class ComponentMap {
       if (this.mainFile === currentFile.relativePath) this.mainFile = newLocation;
       changes.push({ from: currentFile.relativePath, to: newLocation });
       currentFile.relativePath = newLocation;
+      currentFile.name = path.basename(newLocation);
     }
     this.validate();
     return changes;
@@ -201,7 +202,7 @@ export default class ComponentMap {
         componentPaths: [trackDirRelative || '.'],
         id: id.toString(),
         override: false, // this makes sure to not override existing files of componentMap
-        writeToBitMap: false,
+        trackDirFeature: true,
         origin: this.origin
       };
       const numOfFilesBefore = this.files.length;
@@ -230,6 +231,10 @@ export default class ComponentMap {
       return relativePaths.includes(file.relativePath) ? accumulator : accumulator.concat(file);
     }, []);
     this.validate();
+  }
+
+  sort() {
+    this.files = R.sortBy(R.prop('relativePath'), this.files);
   }
 
   validate(): void {