+++ src/consumer/component/add-components/add-components.js
@@ -24,7 +24,7 @@ import { Consumer } from '../../../consumer';
 import BitMap from '../../../consumer/bit-map';
 import { BitId } from '../../../bit-id';
 import type { BitIdStr } from '../../../bit-id/bit-id';
-import { COMPONENT_ORIGINS, REGEX_PATTERN, DEFAULT_DIST_DIRNAME } from '../../../constants';
+import { COMPONENT_ORIGINS, REGEX_PATTERN, DEFAULT_DIST_DIRNAME, VERSION_DELIMITER } from '../../../constants';
 import logger from '../../../logger/logger';
 import {
   PathsNotExist,
@@ -33,12 +33,14 @@ import {
   NoFiles,
   DuplicateIds,
   EmptyDirectory,
+  TestIsDirectory,
   ExcludedMainFile
 } from './exceptions';
 import type { ComponentMapFile, ComponentOrigin } from '../../bit-map/component-map';
 import type { PathLinux, PathOsBased } from '../../../utils/path';
 import ComponentMap from '../../bit-map/component-map';
 import GeneralError from '../../../error/general-error';
+import VersionShouldBeRemoved from './exceptions/version-should-be-removed';
 
 export type AddResult = { id: string, files: ComponentMapFile[] };
 export type AddActionResults = { addedComponents: AddResult[], warnings: Object };
@@ -90,7 +92,7 @@ export type AddProps = {
   tests?: PathOrDSL[],
   exclude?: PathOrDSL[],
   override: boolean,
-  writeToBitMap?: boolean,
+  trackDirFeature?: boolean,
   origin?: ComponentOrigin
 };
 
@@ -104,6 +106,7 @@ export default class AddComponents {
   tests: PathOrDSL[];
   exclude: PathOrDSL[];
   override: boolean; // (default = false) replace the files array or only add files.
+  trackDirFeature: ?boolean;
   writeToBitMap: boolean; // (default = true)
   warnings: Object;
   ignoreList: string[];
@@ -119,7 +122,8 @@ export default class AddComponents {
     this.tests = addProps.tests || [];
     this.exclude = addProps.exclude || [];
     this.override = addProps.override;
-    this.writeToBitMap = addProps.writeToBitMap !== undefined ? addProps.writeToBitMap : true;
+    this.trackDirFeature = addProps.trackDirFeature;
+    this.writeToBitMap = !addProps.trackDirFeature; // trackDir feature doesn't need to write to bitMap
     this.origin = addProps.origin || COMPONENT_ORIGINS.AUTHORED;
     this.warnings = {};
   }
@@ -151,14 +155,20 @@ export default class AddComponents {
   }
 
   addToBitMap({ componentId, files, mainFile, trackDir }: AddedComponent): AddResult {
-    const componentMap: ComponentMap = this.bitMap.addComponent({
-      componentId,
-      files,
-      mainFile,
-      trackDir,
-      origin: COMPONENT_ORIGINS.AUTHORED,
-      override: this.override
-    });
+    const getComponentMap = (): ComponentMap => {
+      if (this.trackDirFeature) {
+        return this.bitMap.addFilesToComponent({ componentId, files });
+      }
+      return this.bitMap.addComponent({
+        componentId,
+        files,
+        mainFile,
+        trackDir,
+        origin: COMPONENT_ORIGINS.AUTHORED,
+        override: this.override
+      });
+    };
+    const componentMap = getComponentMap();
     return { id: componentId.toString(), files: componentMap.files };
   }
 
@@ -196,7 +206,7 @@ export default class AddComponents {
           if (idOfFileIsDifferent) {
             const existingIdWithoutVersion = BitId.parse(existingIdOfFile).toStringWithoutVersion();
             // $FlowFixMe $this.id is not null at this point
-            throw new IncorrectIdForImportedComponent(existingIdWithoutVersion, this.id);
+            throw new IncorrectIdForImportedComponent(existingIdWithoutVersion, this.id, file.relativePath);
           }
         } else if (idOfFileIsDifferent) {
           // not imported component file but exists in bitmap
@@ -241,7 +251,17 @@ export default class AddComponents {
     If you're trying to add a new component, please choose a new namespace or name.
     If you're trying to update a dependency component, please re-import it individually`);
     }
-
+    if (currentId.includes(VERSION_DELIMITER)) {
+      if (
+        !existingComponentId || // this id is new, it shouldn't have a version
+        !existingComponentId.includes(VERSION_DELIMITER) || // this component is new, it shouldn't have a version
+        // user shouldn't add files to a an existing component with different version
+        BitId.getVersionOnlyFromString(existingComponentId) !== BitId.getVersionOnlyFromString(this.id)
+      ) {
+        // $FlowFixMe this.id is defined here
+        throw new VersionShouldBeRemoved(this.id);
+      }
+    }
     return existingComponentId ? BitId.parse(existingComponentId) : BitId.parse(currentId);
   }
 
@@ -293,14 +313,18 @@ export default class AddComponents {
   }
 
   async _mergeTestFilesWithFiles(files: ComponentMapFile[]): Promise<ComponentMapFile[]> {
-    const testFilesArr = !R.isEmpty(this.tests)
+    const testFiles = !R.isEmpty(this.tests)
       ? await this.getFilesAccordingToDsl(files.map(file => file.relativePath), this.tests)
       : [];
-    const resolvedTestFiles = testFilesArr.map(testFile => ({
-      relativePath: testFile,
-      test: true,
-      name: path.basename(testFile)
-    }));
+
+    const resolvedTestFiles = testFiles.map((testFile) => {
+      if (isDir(path.join(this.consumer.getPath(), testFile))) throw new TestIsDirectory(testFile);
+      return {
+        relativePath: testFile,
+        test: true,
+        name: path.basename(testFile)
+      };
+    });
 
     return unionBy(resolvedTestFiles, files, 'relativePath');
   }